Fixed couple of errors
This optimization is based on the assumption that only one VIP is configured
and one real is set for the connection.
Multiple reals can also be configured, but what we do is to offload the configuration
directly into the code, removing all the ipv6 support and chosing the position of the
real based on the configuration set at startup.